edited delete to be more secure
people cannot just delete random people's profiles

+++ controllers/api/user-routes.js
@@ -1,6 +1,8 @@
 const router = require("express").Router();
+const withAuth = require('../utils/auth');
 const { Status, User } = require("../../models");
 const bcrypt = require("bcrypt");
+
 const SALT_FACTOR = 10
 
 // the application end point is /api/user
@@ -91,6 +93,7 @@ router.post('/login', async (req,res) => {
             req.body.password,
             userData.password
         );
+
         // if the passwords do not match, login fails and route ends
         if (!validPassword) {
             res
@@ -99,12 +102,15 @@ router.post('/login', async (req,res) => {
             return;
         }
       
-            req.session.save(() => {
-                req.session.user_id = userData.id;
-                req.session.logged_in = true;
+        req.session.user_id = userData.id;
+        req.session.logged_in = true;
             
-                res.json({ user: userData, message: 'You are now logged in!' });
-            });
+        res.json({ user: userData, message: 'You are now logged in!' });
+
+        // THESE ARE USED FOR TESTING. DELETE AT THE END !!
+        // console.log(req.session.user_id)
+        // console.log(req.session.logged_in)
+    
     } catch (err) {
         res.status(500).json(err)
     }
@@ -130,20 +136,49 @@ router.put("/:id", async (req, res) => {
     }
 })
 
-// deletes user by id 
-router.delete("/:id", async (req, res) => {
-    try {
-        const userData = await User.destroy({
-            where: {
-                id: req.params.id
-            }
+// gets rid of session when logging out
+router.post('/logout', (req, res) => {
+    if (req.session.logged_in) {
+        req.session.destroy(() => {
+            res.status(204).end();
         });
-        if(!userData){
-            res.status(404).json({
-                message: "No user associated with that id"
-            })
-            return;
+    } else {
+        res.status(404).end();
+    }
+});
+
+// // deletes user by id 
+// router.delete("/:id", async (req, res) => {
+//     try {
+        
+//         const userData = await User.destroy({
+//             where: {
+//                 id: req.params.id
+//             }
+//         });
+//         if(!userData){
+//             res.status(404).json({
+//                 message: "No user associated with that id"
+//             })
+//             return;
+//         }
+//         res.status(200).json(userData)
+//     } catch (err) {
+//         res.status(500).json(err)
+//     }
+// });
+
+// deletes user by id // STILL NEEDS TO BE TESTED
+router.delete("/delete", withAuth, async (req, res) => {
+    try {
+        if (req.session.logged_in) {
+            const userData = await User.destroy({
+                where: {
+                    id: req.session.user_id
+                }
+            });
         }
+
         res.status(200).json(userData)
     } catch (err) {
         res.status(500).json(err)
